refactor: AuthorizationServerConfig 관심사 분리
