+++ config/src/main/resources/shared/gateway.yml
@@ -25,7 +25,7 @@ spring:
         - id: auth-service
           uri: lb://auth-service
           predicates:
-            - Path=/oauth2/**,/api/v1/users/**
+            - Path=/oauth2/**,/api/v1/users/**,'.well-known/**'
           filters:
             - StripPrefix=0
